* Fixed viewWrite() and viewRead(). * Replaced CViewManager::Counts type to uint16_t.

+++ llc/gpk_view_manager.h
@@ -34,22 +34,22 @@ namespace gpk
 
 	template<typename _tElement, size_t _blockSize>
 	class CViewManager	{
-		static constexpr	const uint32_t												BLOCK_SIZE					= _blockSize;
+		static constexpr	const uint32_t												BLOCK_SIZE					= (uint32_t)_blockSize;
 							::gpk::unshrinkable_view_container<_tElement, BLOCK_SIZE>	Elements;
 	public:
-							::gpk::array_pod<uint32_t>									Counts;
+							::gpk::array_pod<uint16_t>									Counts;
 							::gpk::array_pod<const _tElement*>							Views;
 
 																						~CViewManager				()																					{}
 																						CViewManager				()																					{ Views.push_back(0); Counts.push_back(0); }
 
-			inline			::gpk::error_t												View						(const _tElement* elements, uint32_t count)													{ ::gpk::view_array<const _tElement> out_view; return View(elements, count, out_view); }
-							::gpk::error_t												View						(const _tElement* elements, uint32_t count, ::gpk::view_array<const _tElement>& out_view)	{
+			inline			::gpk::error_t												View						(const _tElement* elements, uint16_t count)													{ ::gpk::view_array<const _tElement> out_view; return View(elements, count, out_view); }
+							::gpk::error_t												View						(const _tElement* elements, uint16_t count, ::gpk::view_array<const _tElement>& out_view)	{
 			if(0 == count || 0 == elements) {
 				out_view																		= {};
 				return 0;
 			}
-			const uint32_t																		totalChars					= ::gpk::min(count, CViewManager::BLOCK_SIZE);
+			const uint32_t																		totalChars					= ::gpk::min((uint32_t)count, CViewManager::BLOCK_SIZE);
 			for(uint32_t iView = 0, countLabels = Views.size(); iView < countLabels; ++iView) {
 				if(totalChars != Counts[iView])
 					continue;
@@ -60,7 +60,7 @@ namespace gpk
 				}
 			}
 			gpk_necall(Elements.push_sequence(elements, totalChars, out_view)	, "%s", "Out of memory?");
-			const uint32_t																		newIndex					= Counts.push_back(out_view.size());
+			const uint32_t																		newIndex					= Counts.push_back((uint16_t)out_view.size());
 			gpk_necall(newIndex, "%s", "Out of memory?");
 			gpk_necall(Views.push_back(out_view.begin())					, "%s", "Out of memory?");
 			return newIndex;
